+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 0.0.2 (2022-11-11)
++ implemented latitudinal distribution at any taxa level using both `matplotlib` and `plotly`
+
 ## 0.0.1 (2022-11-03)
 + setting up package
-+ implemented taxonomic distribution sunburst plot using plotly
++ implemented taxonomic distribution sunburst plot using `plotly`